feat: 이미지 서치 커스텀훅 분리
